Change name authClients and legacyAuthConsumers
